% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptransp_download.R
\name{ptransp_download}
\alias{ptransp_download}
\title{Download Portal}
\usage{
ptransp_download(opendata, reference, destfile, filename, ...)
}
\description{
This function allows to download the set of data available in the
\url{http://www.portaltransparencia.gov.br}
}
\examples{
ptransp_download(opendata = 1,reference = 2018)
ptransp_download(opendata = "1",reference = "2018")
ptransp_download(opendata = "orcamento-despesa",reference = "2018")
ptransp_download(opendata = 1,reference = 2018,destfile = "~/")
ptransp_download(opendata = 1,reference = 2018,destfile = "~/data",filename = "orcamento.zip")
}
