% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/potr_download_last.R
\name{potr_download_last}
\alias{potr_download_last}
\title{Download the last files of the 'Portal da Transparencia'}
\usage{
potr_download_last(opendata, destfile, filename, download.file.mode, ...)
}
\arguments{
\item{opendata}{A character or numeric thats indicates a dataset selected for download.
See x to meet valid input of the argument.}

\item{destfile}{A character string with the name where the downloaded file is saved.
If missing, the file will be saved to the current directory.}

\item{filename}{File name to dataset. For all dataset defined by 'opendata' use the
extension '.zip' to the end. If missing, the file name will be a title of opendata with a reference.}

\item{download.file.mode}{The mode with which to write the file. If download.file.mode is not found
and the operating system is windows, it will use the "wb".}

\item{...}{Others arguments of the \code{\link[utils:download.file]{download.file}} function.}
}
\description{
This function can be used to download the latest data set from the 'Transparency Portal'
of Brazil - \url{http://www.portaltransparencia.gov.br}.
}
\details{
The function downloads files available in the repository of transparency portal -
website of the federal government of Brazil on public transparency, with budgetary
data and government management. The data disclosed in the Portal come from various
sources of government information. See http://www.portaltransparencia.gov.br/ for
more information.

The function arguments are specifications about the data set to be downloaded.
'opendata' is related to the content of the data, 'reference' to the date of
availability of the file in the portal and 'destfile' to the location to be saved.
Other '...' arguments are related to the download.file function of the 'utils' package.

It is suggested that the values for '...' be reported only if the default setting
does not work. See \code{\link[utils:download.file]{download.file}} for more
information about these arguments and proxy settings.
}
\examples{
potr_download_last(opendata = 19)
potr_download_last(opendata = 19, destfile = "~/data")
potr_download_last(opendata = 19, destfile = "~/data", filename = "orcamento.zip")
potr_download_last(opendata = "19")
potr_download_last(opendata = "19", destfile = "~/data")
potr_download_last(opendata = "19", destfile = "~/data", filename = "orcamento.zip")
potr_download_last(opendata = "orcamento-despesa")
potr_download_last(opendata = "orcamento-despesa", destfile = "~/data")
potr_download_last(opendata = "orcamento-despesa", destfile = "~/data", filename = "orcamento.zip")
}
